requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
requests==2.31.0
websockets==12.0
selenium==4.15.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
